arm-virt: fix final-review findings (timer tick, stack sizing, FIRST_BOOT, docs)
